refactor ExportFromArchive and Request classes

+++ src/main/java/ru/siblion/nesterov/logreader/util/ExportFromArchive.java
@@ -5,6 +5,7 @@ import ru.siblion.nesterov.logreader.type.Config;
 import javax.ejb.Singleton;
 import javax.ejb.Startup;
 import java.io.File;
+import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -26,35 +27,35 @@ public class ExportFromArchive {
 
     private static final Logger logger = MyLogger.getLogger();
 
-    static public void exportResource(String resourceName, String copyFilePath) throws Exception {
+    static public void exportResource(String resourceName, String copyFilePath) {
         String resource = resourceName.replace('\\', '/'); // метод getResourceAsStream использует '/' для разделения в пути файла
         InputStream file = ExportFromArchive.class.getResourceAsStream(resource);
-        Files.copy(file, Paths.get(copyFilePath));
+        try {
+            File checkingFile = new File(copyFilePath); // TODO: 06.02.2017 Нужен ли checkingFile 
+            if (!checkingFile.exists()) {
+                Files.copy(file, Paths.get(copyFilePath));
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     static public void exportConfig() {
         new File(DOMAIN_DIRECTORY + "\\logreader\\config").mkdirs();
-        try {
-            exportResource("\\config\\config.xml", DIRECTORY + "\\config\\config.xml");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        exportResource("\\config\\config.xml", DIRECTORY + "\\config\\config.xml");
     }
 
     static public void exportXsls() {
         config = Config.getConfig(configFile);
         new File(DOMAIN_DIRECTORY + "\\logreader\\xsl").mkdirs();
         new File(DOMAIN_DIRECTORY + "\\logreader\\logs").mkdirs();
-        try {
-            exportResource("\\xsl\\doc.xsl", DIRECTORY + "\\xsl\\doc.xsl");
-            exportResource("\\xsl\\html.xsl", DIRECTORY + "\\xsl\\html.xsl");
-            exportResource("\\xsl\\pdf.xsl", DIRECTORY + "\\xsl\\pdf.xsl");
-            exportResource("\\xsl\\rtf.xsl", DIRECTORY + "\\xsl\\rtf.xsl");
 
-            //exportResource("\\xsl\\siblion_logo.gif", DIRECTORY + "\\xsl\\siblion_logo.gif");;
-            exportResource("\\xsl\\siblion_logo.gif", config.getDirectory() + "\\siblion_logo.gif");
-        } catch (Exception e) {
-            logger.log(Level.SEVERE, "Ошибка при экспорте из архива ", e) ;
-        }
+        exportResource("\\xsl\\doc.xsl", DIRECTORY + "\\xsl\\doc.xsl");
+        exportResource("\\xsl\\html.xsl", DIRECTORY + "\\xsl\\html.xsl");
+        exportResource("\\xsl\\pdf.xsl", DIRECTORY + "\\xsl\\pdf.xsl");
+        exportResource("\\xsl\\rtf.xsl", DIRECTORY + "\\xsl\\rtf.xsl");
+
+        //exportResource("\\xsl\\siblion_logo.gif", DIRECTORY + "\\xsl\\siblion_logo.gif");;
+        exportResource("\\xsl\\siblion_logo.gif", config.getDirectory() + "\\siblion_logo.gif");
     }
 }
